requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44